the problems of tasks and tasks list has been solved

+++ backend/controllers/tasks.controller.js
@@ -1,8 +1,26 @@
-const Tasks     = require('../models/tasks.model');
+const Tasks      = require('../models/tasks.model');
+const TasksList  = require('../models/tasks-list.model');
+const Boards     = require('../models/board');
+const Workspaces = require('../models/spaceWork');
 
 const create = async(req, res) => {
     try {
         if( !req.body.name || !req.body.list_id || !req.body.priority ) return res.status(400).send('Incomplete data');
+
+        const list = await TasksList.findById(req.body.list_id);
+
+        if( !list ) return res.status(400).send('Enter a valid task list');
+
+        const board = await Boards.findById( list.board_id );
+
+        if( !board ) return res.status(400).send('Enter a valid board');
+
+        const space = await Workspaces.findOne({
+            user_id: req.user._id,
+            _id: board.workspace_id,
+        });
+
+        if( !space ) return res.status(400).send('Enter a valid board');
 
         const priority = parseInt( req.body.priority );
 
@@ -13,6 +31,7 @@ const create = async(req, res) => {
             description : req.body.description,
             is_archived : false,
             list_id     : req.body.list_id,
+            priority,
         });
 
         const result = await task.save();
@@ -28,28 +47,66 @@ const create = async(req, res) => {
 }
 
 const list = async(req, res) => {
-    const tasks = await Tasks.find({ list_id: req.params.list_id });
-    return res.status(200).send({ data: tasks });
+    try {
+        const list = await TasksList.findById(req.params.list_id);
+
+        if( !list ) return res.status(400).send('Enter a valid task list');
+
+        const board = await Boards.findById( list.board_id );
+
+        if( !board ) return res.status(400).send('Enter a valid board');
+
+        const space = await Workspaces.findOne({
+            user_id: req.user._id,
+            _id: board.workspace_id,
+        });
+
+        if( !space ) return res.status(400).send('Enter a valid board');
+
+        const tasks = await Tasks.find({ list_id: req.params.list_id });
+        return res.status(200).send({ data: tasks });
+
+    } catch (e) {
+        console.log(`tasks controller create error: ${e}`);
+        return res.status(400).send('An error ocurred listing tasks');
+    }
 }
 
 const update = async(req, res) => {
     try {
-        if( !req.body._id || !req.body.name || !req.body.list_id || !req.body.is_archived || !req.body.priority ) return res.status(400).send('Incomplete data');
+        if( !req.body._id || !req.body.name || !req.body.list_id || !req.body.priority ) return res.status(400).send('Incomplete data');
+
+        const list = await TasksList.findById(req.body.list_id);
+
+        if( !list ) return res.status(400).send('Enter a valid task list');
+
+        const board = await Boards.findById( list.board_id );
+
+        if( !board ) return res.status(400).send('Enter a valid board');
+
+        const space = await Workspaces.findOne({
+            user_id: req.user._id,
+            _id: board.workspace_id,
+        });
+
+        if( !space ) return res.status(400).send('Enter a valid board');
 
         const priority = parseInt( req.body.priority );
 
         if( priority > 5 || priority < 0 ) return res.status(400).send('Enter a valid priority');
 
-        const result = await Tasks.findByIdAndUpdate(req.body._id, {
+        const task = await Tasks.findByIdAndUpdate(req.body._id, {
             name        : req.body.name,
             description : req.body.description,
             is_archived : req.body.is_archived,
             list_id     : req.body.list_id,
         });
 
+        const result = await task.save();
+
         if( !result ) return res.status(400).send('An error ocurred updating task');
 
-        return res.status(200).send({ data: task });
+        return res.status(200).send({ data: task._id });
 
     } catch (e) {
         console.log(`tasks controller update error: ${e}`);
@@ -59,6 +116,24 @@ const update = async(req, res) => {
 
 const del = async(req, res) => {
     try {
+        const task = await Tasks.findById(req.params._id);
+
+        if( !task ) return res.status(401).send('Enter a valid task');
+
+        const list = await TasksList.findById(task.list_id);
+
+        if( !list ) return res.status(400).send('Enter a valid task list');
+
+        const board = await Boards.findById( list.board_id );
+
+        if( !board ) return res.status(400).send('Enter a valid board');
+
+        const space = await Workspaces.findOne({
+            user_id: req.user._id,
+            _id: board.workspace_id,
+        });
+
+        if( !space ) return res.status(400).send('Enter a valid board');
 
         const result = await Tasks.findByIdAndDelete(req.params._id);
